requirements.txt
pyyaml
python-box
numpy
pandas
matplotlib
comet-ml
torch
scikit-learn